+++ backend/src/main/java/backend/capstone/domain/kakaoplace/config/KakaoLocalConfig.java
@@ -1,6 +1,7 @@
 package backend.capstone.domain.kakaoplace.config;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
@@ -15,9 +16,9 @@ public class KakaoLocalConfig {
     @Value("${kakao.local.rest-api-key}")
     private String restApiKey;
 
-    @Bean
+    @Bean("kakaoLocalWebClient")
     public WebClient kakaoLocalWebClient(
-        WebClient.Builder builder
+        @Qualifier("webClientBuilder") WebClient.Builder builder
     ) {
         return builder
             .baseUrl(baseUrl)
