+++ backend/src/main/java/backend/capstone/domain/kakaoplace/service/KakaoSearchByKeywordService.java
@@ -6,6 +6,7 @@ import backend.capstone.domain.kakaoplace.mapper.KakaoPlaceMapper;
 import backend.capstone.domain.kakaoplace.service.dto.KakaoSearchByKeywordResult;
 import backend.capstone.global.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -15,7 +16,7 @@ import org.springframework.web.reactive.function.client.WebClientException;
 @RequiredArgsConstructor
 public class KakaoSearchByKeywordService {
 
-    private final WebClient kakaoLocalWebClient;
+    private final @Qualifier("kakaoLocalWebClient") WebClient kakaoLocalWebClient;
 
     @Transactional(readOnly = true)
     public PlaceSearchResponse searchByKeyword(String query, int page) {
